Changed to percentage cursor values

+++ src/mouse.h
@@ -3,8 +3,9 @@
 
 #define WINVER 0x0500
 #include <windows.h>
+#include <math.h>
 
-void moveMouse(int, int);
+void moveMouse(double, double);
 void clickMouse();
 
 #endif
